UpdateItemController Actions to consider User

+++ WishList/Controllers/ItemController.cs
@@ -1,12 +1,17 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
         public ItemController(ApplicationDbContext context)
         {
@@ -15,8 +20,8 @@ namespace WishList.Controllers
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
-
+            var User = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items.Where(e=>e.User.Id==User.Id).ToList();
             return View("Index", model);
         }
 
@@ -36,7 +41,10 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+            var User = _userManager.GetUserAsync(HttpContext.User).Result;
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if (item.User != User)
+                return Unauthorized();
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
